# Automatic table sizing for HyperClockCache (AutoHCC)